# backend/boxen/assets/openapi.yaml
openapi: 3.1.2
...
info:
  title: Boxen Local API
  version: 1.0.0
  summary: Same-origin API for the locally hosted Boxen inventory system.
  description: >-
    This contract is normative for Boxen v1 HTTP field names and wire formats.
    The API is served only from the configured local Boxen origin: trusted-LAN
    HTTP is supported in development; production requires HTTPS.
    Anonymous access is configured server-side as off, viewer, or editor (default).
    When enabled, GET /session issues a per-browser session cookie and CSRF token
    without login. Other protected operations still require that session cookie.
    Anonymous access never grants user management, system administration, backup,
    or permanent purge permissions.
servers:
  - url: /api/v1
    description: Same-origin local Boxen server
...
security:
  - cookieSession: []
paths:
...
  /boxes/{box_code}/label.pdf:
    get:
      tags: [Labels]
      operationId: renderBoxLabel
      summary: Render a print-ready label PDF. Editor or owner.
      parameters:
        - $ref: '#/components/parameters/BoxCode'
        - name: profile
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 80 }
      responses:
        '200':
          description: Deterministic label PDF.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Content-Disposition:
              schema: { type: string }
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: boxen_session
      description: >-
        HttpOnly, SameSite=Strict session cookie, Secure over HTTPS. Issued by
        local login, first-owner setup, or GET /session when anonymous access is
        enabled. Anonymous mutations retain the same Origin, CSRF, role, and
        resource-version checks as local user sessions.
  headers:
...
    ETag:
      description: Strong resource version identifier.
      schema: { type: string }
  parameters:
...
    BoxCode:
      name: box_code
      in: path
      required: true
      schema: { $ref: '#/components/schemas/BoxCode' }
...
  responses:
...
    Unauthorized:
      description: Authentication is required or no longer valid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Forbidden:
      description: The current user is not authorized.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    ValidationFailed:
      description: One or more request fields failed validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
...
    BoxCode:
      type: string
      pattern: '^BX-[0-9A-HJKMNP-TV-Z]{4}-[0-9A-HJKMNP-TV-Z]{4}$'
      examples: [BX-7K3M-R9QA]
...
    FieldError:
      type: object
      additionalProperties: false
      required: [path, code, message]
      properties:
        path: { type: string }
        code: { type: string }
        message: { type: string }
    ProblemDetails:
      type: object
      additionalProperties: true
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
        request_id: { type: string }
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
...